requirements.txt
requests==2.27.1
fastapi==0.70.1
uvicorn==0.16.0
pydantic==1.9.0
beautifulsoup4~=4.9.0
lxml~=4.6.3
python-docx==0.8.11
